requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
async-timeout==5.0.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
clarabel==0.9.0
click==8.1.8
contourpy==1.2.1
cvxpy==1.5.3
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
ecos==2.0.14
einops==0.8.0
eval_type_backport==0.2.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imageio==2.35.1
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==1.24.4
openai==1.82.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
osqp==0.6.7.post1
pandas==2.2.2
pillow==10.4.0
propcache==0.2.1
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qdldl==0.1.7.post4
regex==2024.7.24
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
scs==3.2.7
shellingham==1.5.4
sniffio==1.3.1
supervision==0.23.0
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.7.0
together==1.3.10
tokenizers==0.20.0
tqdm==4.66.5
typer==0.15.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.18.3

# Qwen requirements
packaging==23.1
ninja==1.11.1.2
transformers==4.46.3
accelerate==1.0.1
qwen-vl-utils
ray==2.40.0
torchvision
starlette==0.41.3
fastapi==0.115.6
uvicorn==0.34.0